Reorganize modules (Django project and app) (#2)
Mostly renaming folders. The main project is renamed from "Macon_Code" to "config" and brought up to the root level of the repo. I was inspired by this discussion and liked the "config" idea.

Changes of note:

Bootstrap packages have been removed as dependencies and included directly in the template html for now.
Static files and templates are now in a subfolder matching the app name, as recommended in the Django tutorials.